% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{GetPopulationPyramidKSH}
\alias{GetPopulationPyramidKSH}
\title{Scrape the population pyramid of Hungary (2015-2017).}
\usage{
GetPopulationPyramidKSH(Type = "Jan1", Years = 2015:2017,
  Gender = "Total", AgeGroup = "FiveYear", GeographicArea = "NUTS2")
}
\arguments{
\item{Type}{Type of population (possible values are "Jan1" and "MidYear").}

\item{Years}{Numeric vector containing the years (from 2015 to 2017, inclusive).}

\item{Gender}{Stratification according to gender (possible values are "Total" and "Both").}

\item{AgeGroup}{Stratification according to age (possible values are "Total", "Main" and
"FiveYear").}

\item{GeographicArea}{Stratification according to geographic area (possible values are "NUTS2",
"NUTS3" and "LAU1").}
}
\value{
Neatly formatted population pyramid.
}
\description{
Scrape the population pyramid of Hungary (2015-2017) possibly stratified according to gender,
age group (main groups or five-year groups) and geographic area (NUTS2, NUTS3 or LAU1).
}
\examples{
GetPopulationPyramidKSH( "MidYear", Years = 2015, AgeGroup = "Main" )
}
