% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{KSHStatinfoScrape}
\alias{KSHStatinfoScrape}
\title{Scrape the Statinfo dissemination database.}
\usage{
KSHStatinfoScrape(code, column, row)
}
\arguments{
\item{code}{The code of the data set (given under the label 'Technical identifier').}

\item{column}{A list of vectors, where each vectors is a stratificiation variable for the column,
with the elements of the vector specifying the strata.}

\item{row}{A list of vectors, where each vectors is a stratificiation variable for the row, with
the elements of the vector specifying the strata.}
}
\value{
The raw file returned by the Statinfo database as a data frame.
}
\description{
Scrape the Statinfo dissemination database of the Hungarian Central Statistics Office.
}
\examples{
KSHStatinfoScrape( "NT5C01", "[NTAC001]", list( c( "[2015]", "[2016]" ), c( "[N101_1]", "[N101_2]" ) ) )
}
